# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для пользователей, мест, отзывов и токенов",
        "title": "Reviewlink API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/leaderboard/bonuses": {
            "get": {
                "description": "Возвращает список пользователей, отсортированных по количеству полученных бонусов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Получить топ пользователей по бонусам",
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BonusLeaderboardEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/leaderboard/places": {
            "get": {
                "description": "Возвращает список заведений, отсортированных по количеству отзывов и среднему рейтингу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Получить топ заведений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Максимальное количество результатов (по умолчанию 10, максимум 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка: 'reviews' или 'rating' (по умолчанию 'reviews')",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Минимальный средний рейтинг (по умолчанию 0)",
                        "name": "min_rating",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Минимальное количество отзывов (по умолчанию 0)",
                        "name": "min_reviews",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заведений",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.LeaderboardEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/leaderboard/users": {
            "get": {
                "description": "Возвращает список пользователей, отсортированных по количеству отзывов и среднему рейтингу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Получить топ пользователей",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Максимальное количество результатов (по умолчанию 10, максимум 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка: 'reviews' или 'rating' (по умолчанию 'reviews')",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Минимальный средний рейтинг (по умолчанию 0)",
                        "name": "min_rating",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Минимальное количество отзывов (по умолчанию 0)",
                        "name": "min_reviews",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.LeaderboardEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BonusLeaderboardEntry": {
            "type": "object",
            "properties": {
                "bonuses_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "points_spent": {
                    "type": "integer"
                },
                "rank": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "avg_rating": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "reviews_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}